Fixing 0 quantity inventory

+++ src/main/java/br/com/andre/msscbeerinventoryservice/services/impl/AllocationServiceImpl.java
@@ -75,6 +75,9 @@ public class AllocationServiceImpl implements AllocationService {
                 beerOrderLine.setQuantityAllocated(allocatedQty + inventory);
                 beerInventory.setQuantityOnHand(0);
 
+            }
+
+            if (beerInventory.getQuantityOnHand() == 0) {
                 beerInventoryRepository.delete(beerInventory);
             }
         });
